move sample instructions into its own component
