+++ src/frontend/src/common/components/library/data_subview/components/CreateNSTreeModal/components/SampleIdInput/components/InputInstructions/style.ts
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+import { getFontWeights } from "czifui";
+
+export const SemiBold = styled.span`
+  ${(props) => {
+    const fontWeights = getFontWeights(props);
+    return `
+      font-weight: ${fontWeights?.semibold};
+    `;
+  }}
+`;
